split out large burden estimte controller into 2 classes
