+++ src/app/src/test/kotlin/org/vaccineimpact/api/tests/controllers/BurdenEstimates/UploadBurdenEstimatesControllerTests.kt
@@ -1,0 +1,176 @@
+package org.vaccineimpact.api.tests.controllers.BurdenEstimates
+
+import com.nhaarman.mockito_kotlin.*
+import org.assertj.core.api.Assertions
+import org.junit.Test
+import org.mockito.Mockito
+import org.vaccineimpact.api.app.context.ActionContext
+import org.vaccineimpact.api.app.controllers.BurdenEstimates.BurdenEstimateUploadController
+import org.vaccineimpact.api.app.errors.MissingRowsError
+import org.vaccineimpact.api.app.logic.BurdenEstimateLogic
+import org.vaccineimpact.api.app.repositories.BurdenEstimateRepository
+import org.vaccineimpact.api.app.repositories.Repositories
+import org.vaccineimpact.api.app.repositories.SimpleDataSet
+import org.vaccineimpact.api.app.requests.PostDataHelper
+import org.vaccineimpact.api.models.*
+import org.vaccineimpact.api.tests.mocks.mockCSVPostData
+
+class UploadBurdenEstimatesControllerTests : BurdenEstimateControllerTestsBase()
+{
+
+    @Test
+    fun `can populate central estimate set`()
+    {
+        val touchstoneSet = mockTouchstones()
+        val logic = mockLogic()
+
+        val expectedData = listOf(
+                BurdenEstimateWithRunId("yf", null, 2000, 50, "AFG", "Afghanistan", 1000F, mapOf(
+                        "deaths" to 10F,
+                        "cases" to 100F
+                )),
+                BurdenEstimateWithRunId("yf", null, 1980, 30, "AGO", "Angola", 2000F, mapOf(
+                        "deaths" to 20F,
+                        "dalys" to 73.6F
+                ))
+        )
+
+        val mockContext = mockActionContext()
+        verifyLogicIsInvokedToPopulateSet(mockContext, mockEstimatesRepository(touchstoneSet), logic, touchstoneSet,
+                normalCSVData.asSequence(), expectedData)
+    }
+
+    @Test
+    fun `can populate stochastic estimate set`()
+    {
+        val touchstoneSet = mockTouchstones()
+        val logic = mockLogic()
+
+        val csvData = listOf(
+                StochasticBurdenEstimate("yf", "runA", 2000, 50, "AFG", "Afghanistan", 1000F, mapOf(
+                        "deaths" to 10F,
+                        "cases" to 100F
+                )),
+                StochasticBurdenEstimate("yf", "runB", 1980, 30, "AGO", "Angola", 2000F, mapOf(
+                        "deaths" to 20F,
+                        "dalys" to 73.6F
+                ))
+        )
+        val expectedData = listOf(
+                BurdenEstimateWithRunId("yf", "runA", 2000, 50, "AFG", "Afghanistan", 1000F, mapOf(
+                        "deaths" to 10F,
+                        "cases" to 100F
+                )),
+                BurdenEstimateWithRunId("yf", "runB", 1980, 30, "AGO", "Angola", 2000F, mapOf(
+                        "deaths" to 20F,
+                        "dalys" to 73.6F
+                ))
+        )
+
+        val mockContext = mockActionContext()
+        val repo = mockEstimatesRepository(touchstoneSet, existingBurdenEstimateSet = defaultEstimateSet.copy(
+                type = BurdenEstimateSetType(BurdenEstimateSetTypeCode.STOCHASTIC)))
+        verifyLogicIsInvokedToPopulateSet(mockContext, repo, logic, touchstoneSet,
+                csvData.asSequence(), expectedData)
+    }
+
+    @Test
+    fun `if keepOpen is not provided, populate closes estimate set`()
+    {
+        // This way, the webapps will carry on with the same behaviour as before.
+        // It's only if a client explicitly sets keepOpen that we will see the partial state
+        populateAndCheckIfSetIsClosed(keepOpen = null, expectedClosed = true)
+    }
+
+    @Test
+    fun `if keepOpen is true, burden estimate set is left open`()
+    {
+        populateAndCheckIfSetIsClosed(keepOpen = "true", expectedClosed = false)
+    }
+
+    @Test
+    fun `if keepOpen is false, populate closes burden estimate set`()
+    {
+        populateAndCheckIfSetIsClosed(keepOpen = "false", expectedClosed = true)
+    }
+
+    @Test
+    fun `populate burden estimate set catches missing row error and returns result`()
+    {
+        val logic = mockLogic()
+        Mockito.`when`(logic.closeBurdenEstimateSet(any(), any(), any(), any()))
+                .doThrow(MissingRowsError("message"))
+        val repo = mockEstimatesRepository()
+        val mockContext = mockActionContext()
+        val mockPostData = mockCSVPostData(normalCSVData)
+        val result = BurdenEstimateUploadController(mockContext, mockRepositories(repo), logic, repo,
+                postDataHelper = mockPostData).populateBurdenEstimateSet()
+        Assertions.assertThat(result.status).isEqualTo(ResultStatus.FAILURE)
+    }
+
+    private fun populateAndCheckIfSetIsClosed(keepOpen: String?, expectedClosed: Boolean)
+    {
+        val timesExpected = if (expectedClosed) times(1) else never()
+
+        val logic = mockLogic()
+        val repo = mockEstimatesRepository()
+        val mockContext = mockActionContext(keepOpen = keepOpen)
+        val mockPostData = mockCSVPostData(normalCSVData)
+        BurdenEstimateUploadController(mockContext, mockRepositories(repo), logic, repo, postDataHelper = mockPostData).populateBurdenEstimateSet()
+        verify(logic, timesExpected).closeBurdenEstimateSet(defaultEstimateSet.id,
+                "group-1", "touchstone-1", "scenario-1")
+    }
+
+    private fun mockActionContext(keepOpen: String? = null): ActionContext
+    {
+        return mock {
+            on { username } doReturn "username"
+            on { contentType() } doReturn "text/csv"
+            on { params(":set-id") } doReturn "1"
+            on { params(":group-id") } doReturn "group-1"
+            on { params(":touchstone-version-id") } doReturn "touchstone-1"
+            on { params(":scenario-id") } doReturn "scenario-1"
+            on { queryParams("keepOpen") } doReturn keepOpen
+        }
+    }
+
+    private fun <T : Any> verifyLogicIsInvokedToPopulateSet(
+            actionContext: ActionContext,
+            repo: BurdenEstimateRepository,
+            logic: BurdenEstimateLogic,
+            touchstoneVersionSet: SimpleDataSet<TouchstoneVersion, String>,
+            actualData: Sequence<T>,
+            expectedData: List<BurdenEstimateWithRunId>
+    )
+    {
+        val postDataHelper = mock<PostDataHelper> {
+            on { csvData<T>(any(), any()) } doReturn actualData
+        }
+
+        val sut = BurdenEstimateUploadController(actionContext, mockRepositories(repo), logic, repo,
+                postDataHelper = postDataHelper)
+
+        sut.populateBurdenEstimateSet()
+        verify(touchstoneVersionSet).get("touchstone-1")
+        verify(logic).populateBurdenEstimateSet(eq(1),
+                eq("group-1"), eq("touchstone-1"), eq("scenario-1"),
+                argWhere { it.toSet() == expectedData.toSet() }
+        )
+    }
+
+    private val normalCSVData = listOf(
+            BurdenEstimate("yf", 2000, 50, "AFG", "Afghanistan", 1000F, mapOf(
+                    "deaths" to 10F,
+                    "cases" to 100F
+            )),
+            BurdenEstimate("yf", 1980, 30, "AGO", "Angola", 2000F, mapOf(
+                    "deaths" to 20F,
+                    "cases" to 73.6F
+            ))
+    )
+
+    private fun mockRepositories(repo: BurdenEstimateRepository) = mock<Repositories> {
+        on { burdenEstimates } doReturn repo
+    }
+
+}
